+++ game/src/game.cpp
@@ -1,7 +1,8 @@
 #include "../include/game.h"
+#include <iostream>
 
 void moveAndResolveCollision(World &currentWorld, SDL_FRect &target, float &xVel, float &yVel) {
-	int iterations = ceil(deltaTime * 2);
+	Uint64 iterations = static_cast<Uint64>(ceil(deltaTime * 2));
 	if (iterations > 128) iterations = 128;
 	float prevX = target.x;
 	float prevY = target.y;
@@ -9,7 +10,7 @@ void moveAndResolveCollision(World &currentWorld, SDL_FRect &target, float &xVel
 
 	for (int i = 0; i < iterations; i++) {
 		target.x += xVel / iterations;
-		collided = isCollidingWithTile(currentWorld, target, 0.1);
+		collided = isCollidingWithTile(currentWorld, target, 0.1f);
 		if (collided) {
 			target.x = prevX;
 			break;
@@ -19,7 +20,7 @@ void moveAndResolveCollision(World &currentWorld, SDL_FRect &target, float &xVel
 
 	for (int i = 0; i < iterations; i++) {
 		target.y += yVel / iterations;
-		collided = isCollidingWithTile(currentWorld, target, 0.1);
+		collided = isCollidingWithTile(currentWorld, target, 0.1f);
 		if (collided) {
 			target.y = prevY;
 			break;
